thinkHard_takeOff_relax_Land
This file reads a file output from Processing. The MindFlex Headset data
is saved in that file. The if the Attention value is over a benchmark
number, the AR Drone 2.0 will launch. It will stay in the air until the
Attention value drops below the benchmark value.  Remember! If you take
the headset off when the number is high, the drone will NOT LAND BY
DEFAULT. So be careful okay :-D.