3 Headsets
Use 3 mindflex headsets to control the "launch", "rotate", and "land"
commands on the AR Drone 2.0. It's set up like a game, where whichever
headset has the highest Attention reading controls the next action. If
headset 1 is thinking hardest, the drone will launch. If 2 is highest,
it will rotate (if it's launched). And if 3 is highest, the drone will
land.